+++ CHANGELOG.rst
@@ -15,6 +15,23 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.12
+------
+2023-02-24
+
+*Note, that this release introduces breaking changes!*
+
+- Make it easy to use a different PDF library with ``PdfFileProvider`` by
+  adding ``pdf_generator_cls`` and ``pdf_generator_kwargs`` optional arguments
+  to the ``pdf_file`` method. Added ``ReportlabPdfGenerator`` class based on
+  the famous ``reportlab`` library. Default is still ``PdfkitPdfGenerator``.
+  Since ``encoding`` was something specific for ``pdfkit`` library, that
+  it was moved from ``pdf_file`` method to ``PdfkitPdfGenerator``, to which it
+  can be passed in ``pdf_generator_kwargs``. If you have passed the
+  ``encoding`` argument explicitly, make sure to make correspondent changes.
+  Note, that using the new ``ReportlabPdfGenerator`` class could speed-up PDF
+  generation by about 40 times.
+
 0.11.5
 ------
 2023-02-20